refactor: Line, Station 생성자 함수 models 폴더로 이동
