+++ src/models/index.js
@@ -1,0 +1,2 @@
+export { default as Line } from './line.js';
+export { default as Station } from './station.js';
